Add basic support for a default IONHOME value.
I want this feature so that I can make an Ion compiler that knows where to find
its files without needing a wrapper script, although it is kind of limited now
because you cannot specify it relative to the executable.

The IONHOME macro must be defined with quotes around it.